--- src/FluentSetups.SourceGenerator/Models/FNamespace.cs
@@ -1,18 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="FNamespace.cs" company="consolovers">
-//   Copyright (c) daniel bramer 2022 - 2022
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace FluentSetups.SourceGenerator.Models
-{
-   using System;
-
-   public class FNamespace : IDisposable
-   {
-      public void Dispose()
-      {
-         
-      }
-   }
-}
